requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make update-requirements
#

requests
tqdm
lxml
click
pandas
numpy
scipy
bs4
pyyaml
pysocks
google-api-python-client
